+++ bitcoin/chainparams.h
@@ -7,6 +7,11 @@
 #include <common/amount.h>
 #include <stdbool.h>
 
+struct bip32_key_version {
+	u32 bip32_pubkey_version;
+	u32 bip32_privkey_version;
+};
+
 struct chainparams {
 	const char *network_name;
 	const char *bip173_name;
@@ -21,6 +26,9 @@ struct chainparams {
 
 	/* Whether this is a test network or not */
 	const bool testnet;
+
+	/* Version codes for BIP32 extended keys in libwally-core*/
+	const struct bip32_key_version bip32_key_version;
 };
 
 /**